Update README.md
This script is designed to fix mouse cursor and GTK themes for various window managers, including sway, openbox, i3, awesome, and more. It ensures a consistent and visually appealing desktop environment. The script provides two main functionalities: a loop for xrdb fixes and a loop for cursor and GTK theme fixes.